+++ shuup/admin/modules/orders/views/__init__.py
@@ -11,7 +11,9 @@ from .detail import OrderDetailView, OrderSetStatusView
 from .edit import OrderEditView, UpdateAdminCommentView
 from .list import OrderListView
 from .log import NewLogEntryView
-from .payment import OrderCreatePaymentView, OrderSetPaidView
+from .payment import (
+    OrderCreatePaymentView, OrderDeletePaymentView, OrderSetPaidView
+)
 from .refund import OrderCreateFullRefundView, OrderCreateRefundView
 from .shipment import OrderCreateShipmentView, ShipmentDeleteView
 from .status import OrderStatusEditView, OrderStatusListView
@@ -21,5 +23,5 @@ __all__ = [
     "OrderListView", "OrderCreatePaymentView", "OrderCreateFullRefundView",
     "OrderCreateRefundView", "OrderCreateShipmentView", "OrderSetPaidView",
     "OrderSetStatusView", "OrderStatusEditView", "OrderStatusListView", "ShipmentDeleteView",
-    "UpdateAdminCommentView"
+    "UpdateAdminCommentView", "OrderDeletePaymentView"
 ]